Add constructor, getters, toString and JavaDocs
Created the constructor that declares a card contains a Value (The worth
o a card), Suit (If the card is Clubs, Diamonds..) and lastly Rank (Ace,
King, Jack..). Added Getters for access. Added toString so that each
card has its name for example Ace of Hearts. Finally added JavaDocs to
everything for easy reading and convention.